docs: summaries for some (service) classes
