+++ TransactionQL.DesktopApp.Tests/Services/InMemoryAccounts.cs
@@ -3,6 +3,10 @@ using TransactionQL.DesktopApp.Services;
 
 namespace TransactionQL.DesktopApp.Tests.Services;
 
+/// <summary>
+/// Simple, non-updating account selector for ease of use in tests.
+/// Uses the <see cref="FuzzyMatcher" /> by default.
+/// </summary>
 public class InMemoryAccounts : ISelectAccounts
 {
     private readonly IMatchAccounts _matcher;
