+++ TransactionQL.DesktopApp/Controls/FilePathConverter.cs
@@ -6,6 +6,9 @@ using System.IO;
 
 namespace TransactionQL.DesktopApp.Controls
 {
+    /// <summary>
+    /// Converts a file path to just the filename.
+    /// </summary>
     public class FilePathConverter : IValueConverter
     {
         public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
@@ -15,16 +18,21 @@ namespace TransactionQL.DesktopApp.Controls
                 return string.Empty;
             }
 
-            return value is string path && targetType == typeof(string)
-                ? Path.GetFileName(path)
-                : value is Uri uri && targetType == typeof(string)
-                ? (object)Path.GetFileName(uri.AbsolutePath)
-                : throw new InvalidCastException($"{value} ({value.GetType()}) is not a file path.");
+            if (value is string path && targetType == typeof(string))
+            {
+                return Path.GetFileName(path);
+            }
+            else if (value is Uri uri && targetType == typeof(string))
+            {
+                return (object)Path.GetFileName(uri.AbsolutePath);
+            }
+            else
+            {
+                throw new InvalidCastException($"{value} ({value.GetType()}) is not a file path.");
+            }
         }
 
         public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
-        {
-            return BindingOperations.DoNothing;
-        }
+            => BindingOperations.DoNothing;
     }
 }
